Split out builder crate
Having the iam-runtime-rs build its own protobuf files meant that crates
which use it must have `protoc` installed so that the dependency can
compile itself. This is annoying.

So I have moved the build bit out into a separate, unpublished, crate.
The other benefit of this is that the files in `builder/proto/` can be
gitignored.